adding qr ics code
