+++ dev-ops/docker/actions/start.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 #DESCRIPTION: starts the docker container
 
-docker-compose up
+docker-compose up -d
