+++ htdocs/src/Oc/GeoCache/StatisticPicture.php
@@ -7,10 +7,8 @@ namespace Oc\GeoCache;
 
 class StatisticPicture
 {
-    public static function deleteStatisticPicture($userId)
+    public static function deleteStatisticPicture(int $userId): void
     {
-        $userId += 0;
-
         // data changed - delete stat pic of user, if exists - will be recreated on next request
         $imagePath = __DIR__ . '/../../../images/statpics/statpic' . $userId . '.jpg';
         if (file_exists($imagePath)) {
